+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngineSession.kt
@@ -106,6 +106,10 @@ class SystemEngineSession(
         throw UnsupportedOperationException("PDF support is not available in this engine")
     }
 
+    override fun requestPrintContent() {
+        throw UnsupportedOperationException("Print support is not available in this engine")
+    }
+
     /**
      * See [EngineSession.stopLoading]
      */