+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/engine/EngineObserver.kt
@@ -435,6 +435,14 @@ internal class EngineObserver(
         store.dispatch(EngineAction.SaveToPdfExceptionAction(tabId, throwable))
     }
 
+    override fun onPrintFinish() {
+        store.dispatch(EngineAction.PrintContentCompletedAction(tabId))
+    }
+
+    override fun onPrintException(isPrint: Boolean, throwable: Throwable) {
+        store.dispatch(EngineAction.PrintContentExceptionAction(tabId, isPrint, throwable))
+    }
+
     override fun onCheckForFormData(containsFormData: Boolean) {
         store.dispatch(ContentAction.UpdateHasFormDataAction(tabId, containsFormData))
     }